scriptharness.unicode. More files = fewer circular imports
