+++ tests/test_exceptions.py
@@ -15,22 +15,6 @@ from . import UNICODE_STRINGS
 class TestUnicode(unittest.TestCase):
     """Test unicode support.
     """
-    def test_to_unicode(self):
-        """test_exceptions | Verify to_unicode gives a unicode string
-        """
-        for ustring in UNICODE_STRINGS:
-            astring = exceptions.to_unicode(ustring)
-            if six.PY2 and not isinstance(ustring, six.text_type):
-                self.assertEqual(ustring.decode('utf-8'), astring)
-            else:
-                self.assertEqual(ustring, astring)
-
-    def test_to_unicode_exception(self):
-        """test_exceptions | Verify to_unicode of a None object gives None
-        """
-        value = exceptions.to_unicode(None)
-        self.assertTrue(value is None)
-
     def test_exception(self):
         """test_exceptions | Verify ScriptHarnessBaseException works
         """
